+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.4.1
 
 - Fix wording on some English strings
+- Fix issue with Edit Mods dialog not showing all